style: organize create_election command
